Crate depth_cloud

Demonstrates the use of our depth cloud renderer, which will efficiently draw a point cloud using a depth texture and a set of intrinsics.

Usage

Native:

cargo r -p re_renderer --example depth_cloud

Web:

cargo run-wasm --example depth_cloud

Modules

framework 🔒

Example framework

Structs

AlbedoTexture 🔒

DepthTexture 🔒

RenderDepthClouds 🔒

Enums

CameraControl 🔒

Functions

hash

main 🔒

spiral 🔒

Returns (position_in_image, linear_depth).